Move Dealer & Player to new file
