+++ PockerGameApp/PockerGameApp/PokerGame.swift
@@ -36,64 +36,6 @@ enum PokerType {
     }
 }
 
-protocol Playable {
-    var cards: [Card] { get }
-    
-    func receive(cards: [Card])
-}
-
-class Player: Playable {
-    let name: String
-    var cards: [Card] = []
-    
-    static var candidates = [
-        "Eddy", "Jason", "Chez", "Dale", "Selina", "Jed", "Jee", "Ebony", "Sol", "Gucci"
-    ]
-    
-    init(name: String) {
-        self.name = name
-    }
-    
-    init() {
-        self.name = Player.candidates.randomElement()!
-        self.cards = []
-    }
-    
-    func receive(cards: [Card]) {
-        self.cards += cards
-    }
-}
-
-class Dealer: Playable {
-    var deck: CardDeck
-    var cards: [Card]
-    var players: [Player]
-    
-    init(deck: CardDeck, players: [Player], cards: [Card] = []) {
-        self.players = players
-        self.deck = deck
-        self.cards = cards
-    }
-    
-    func deal(numOfcards: Int) throws {
-        deck.shuffle()
-        
-        let allPlayers: [Playable] = players + [self]
-        for player in allPlayers {
-            var drawns = [Card]()
-            for _ in 0..<numOfcards {
-                guard let drawn = deck.draw() else { throw PokerGameError.tooManyPlayer }
-                drawns.append(drawn)
-            }
-            player.receive(cards: drawns)
-        }
-    }
-    
-    func receive(cards: [Card]) {
-        self.cards += cards
-    }
-}
-
 enum PokerGameError: Error {
     case tooManyPlayer
 }
